chore: Update `additional_info` column in `MembersTable` to use JSON type
This commit updates the `additional_info` column in the `MembersTable` to use the JSON type. This change ensures that the column stores JSON data and sets a default value of an empty object (`{}`). The `getSchema()` method in the `MembersTable` class is overridden to set the column type to JSON.

This commit also implemented the UI for Additional Info and set some convensions for how to run the UI from the App Settings

+++ app/src/Controller/ActivitiesController.php
@@ -65,6 +65,9 @@ class ActivitiesController extends AppController
                 },
             ],
         );
+        if (!$activity) {
+            throw new \Cake\Http\Exception\NotFoundException();
+        }
         $this->Authorization->authorize($activity);
         $roles = $this->Activities->Permissions->Roles
             ->find()
@@ -145,6 +148,9 @@ class ActivitiesController extends AppController
     public function edit($id = null)
     {
         $activity = $this->Activities->get($id, contain: []);
+        if (!$activity) {
+            throw new \Cake\Http\Exception\NotFoundException();
+        }
         $this->Authorization->authorize($activity);
         if ($this->request->is(["patch", "post", "put"])) {
             $activity = $this->Activities->patchEntity(
@@ -188,6 +194,9 @@ class ActivitiesController extends AppController
     {
         $this->request->allowMethod(["post", "delete"]);
         $activity = $this->Activities->get($id);
+        if (!$activity) {
+            throw new \Cake\Http\Exception\NotFoundException();
+        }
         $this->Authorization->authorize($activity);
         if ($this->Activities->delete($activity)) {
             $this->Flash->success(
